requirements.txt
alabaster==1.0.0
asgiref==3.11.0
babel==2.18.0
certifi==2026.1.4
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.4
Django==6.0.1
djangorestframework==3.16.1
docutils==0.22.4
idna==3.11
imagesize==1.4.1
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
pycparser==3.0
Pygments==2.19.2
PyMySQL==1.1.2
requests==2.32.5
roman-numerals==4.1.0
snowballstemmer==3.0.1
Sphinx==9.1.0
sphinx_rtd_theme==3.1.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
